Fixed various warnings and errors. * winnls.h required in unzip.cc. * Fixed signed/unsigned warnings. * Use sscanf_s on Windows.

+++ src/Bootil/Types/Math.cpp
@@ -8,7 +8,7 @@ class InitialRandomSeeder
 
 		InitialRandomSeeder()
 		{
-			Bootil::Math::Random::Seed( time(NULL) );
+			Bootil::Math::Random::Seed( static_cast< unsigned int >(time(NULL)) );
 		}
 };
